requirements.txt
elasticsearch-dsl==2.2.0
flask==0.12
gunicorn==19.7.0
gevent==1.2.1
google-cloud-datastore==0.23.0